backend/apikeys_bot/requirements.txt
psycopg2==2.9.5
python-telegram-bot==20.1